+++ arandomness/tests/test_MemEater.py
@@ -1,10 +1,10 @@
 #! /usr/bin/env python
 
-"""Generate an arandom number sequence
+"""Test arandomness' MemEater
 
 Copyright:
-    generator.py  generates an arandom number sequence
-    Copyright (C) 2018  Alex Hyer
+    test_MemEater.py  Test arandomness' MemEater
+    Copyright (C) 2017  Alex Hyer
 
     This program is free software: you can redistribute it and/or modify
     it under the terms of the GNU General Public License as published by
@@ -20,25 +20,17 @@ Copyright:
     along with this program.  If not, see <http://www.gnu.org/licenses/>.
 """
 
-from arandomness.memeater import MemEater
-import psutil
-from time import sleep
+from ..memeater import MemEater
 
 __author__ = 'Alex Hyer'
 __email__ = 'theonehyer@gmail.com'
 __license__ = 'GPLv3'
 __maintainer__ = 'Alex Hyer'
-__status__ = 'Planning'
-__version__ = '0.1.0a2'
+__status__ = 'Production'
+__version__ = '1.0.0'
 
 
-def agenerator():
-    """Arandom number generator"""
+def test_MemEater():
+    """Test arandomness' MemEater"""
 
-    free_mem = psutil.virtual_memory().available
-    mem_24 = 0.24 * free_mem
-    mem_26 = 0.26 * free_mem
-    a = MemEater(int(mem_24) / 8)
-    b = MemEater(int(mem_26) / 8)
-    sleep(5)
-    return (free_mem/1000/1000, psutil.virtual_memory().available/1000/1000)
+    assert MemEater(1024).size() == 1024